requirements.txt
playwright
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
alembic>=1.13
asyncpg
python-dotenv